requirements.txt
tk
customtkinter
pandas
Pillow
tkcalendar
SQLAlchemy
